EIP-712 clear-signing

+++ src_features/provideNFTInformation/cmd_provideNFTInfo.c
@@ -55,7 +55,7 @@ typedef bool verificationAlgo(const cx_ecfp_public_key_t *,
 void handleProvideNFTInformation(uint8_t p1,
                                  uint8_t p2,
                                  const uint8_t *workBuffer,
-                                 uint16_t dataLength,
+                                 uint8_t dataLength,
                                  unsigned int *flags,
                                  unsigned int *tx) {
     UNUSED(p1);
